% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{arena.team}
\alias{arena.team}
\title{A dataset containing arean, NHL team names and lat and long of
arean location}
\format{A data frame with 30 rows and 4 variables:
\describe{
  \item{Areana}{arena name}
  \item{team}{team name and its city}
  \item{lat}{latitude of the arean location}
  \item{long}{longitude of the arean location}
  ...
}}
\source{
\url{http://stat545-ubc.github.io/r_package_data/arena_team.tsv}
}
\usage{
arena.team
}
\description{
A dataset containing arean, NHL team names and lat and long of
arean location
}
\keyword{datasets}

