% Generated by roxygen2 (4.0.2): do not edit by hand
\name{check_date}
\alias{check_date}
\title{Is the date format correct?}
\usage{
check_date(gDate)
}
\arguments{
\item{gDate}{is a string corresponding to game day}
}
\value{
Logical \code{TRUE} if \code{date} is in format YYYY-mm-dd,
otherwise, \code{FALSE}.
}
\description{
This function returns TRUE if the input date is in the format YYYY-mm-dd
}
\examples{
check_date("2014-11-07")
check_date("201-411-07")
}

