% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{gameday}
\alias{gameday}
\alias{gameday-package}
\title{gameday}
\description{
Check NHL team play, and scores
}
\details{
Gameday package would allow users to query live.nhle.com to check if your NHL team
is listed on the team play on a given day. Also can query live.nhle.com to get the
scores for a given day.

Functions included

gday()
This function returns TRUE if your NHL team plays
on a given day and FALSE otherwise.

scores()
This function takes a date input and returns
a data.frame with information about the scores of all games on the given date.
}

