% Generated by roxygen2 (4.0.2): do not edit by hand
\name{info_team}
\alias{info_team}
\title{Team information}
\usage{
info_team(team = "canucks")
}
\arguments{
\item{team}{}
}
\value{
data frame with information about arean, team, lat and long
}
\description{
This function returns the information of a team or a city
}
\details{
You know then problem: You're in your office writing R code and
suddenly your friend asks you the name of the arena of a NHL team.
Before you know it you just wasted 15 minutes browsing the lastest
news on your favorite hockey webpage.
Suffer no more! You can now ask R directly, without tempting yourself
by firing up your web browser.
}
\note{
case in \code{team} is ignored
}
\examples{
info_team()
info_team("Bruins")
}
\keyword{misc}

