% Generated by roxygen2 (4.0.2): do not edit by hand
\name{internet_connection}
\alias{internet_connection}
\title{Is there internet connection?}
\usage{
internet_connection()
}
\value{
Logical \code{TRUE} if there is internet connection,
\code{FALSE} otherwise
}
\description{
This function returns TRUE if there is internet connection, otherwise it
returns FALSE
}
\note{
This is a helper function for gday.
}
\examples{
internet_connection()
}

