% Generated by roxygen2 (4.0.2): do not edit by hand
\name{score}
\alias{score}
\title{View the NHL scores for a specific day.}
\usage{
score(date = Sys.Date())
}
\arguments{
\item{date}{Specified date to check scores. Defaults to today's date}
}
\value{
a data.frame that shows the home and away scores for all teams playing on the date specified.
}
\description{
This function returns the NHL scores for a particular day in a convenient data.frame.
}
\details{
Your friend is certain that the Chicago Blackhawks lost the game on November 22, 2014, but you
are quite sure that they won considerably. Because you are working diligently on an R
assignment, you now can check with ease!
}
\note{
\code{date} format is \code{\%Y-\%m-\%d}, but flexible with format \code{\%Y/\%m/\%d}.
}
\examples{
score(Sys.Date())
gday("2014-11-22")
gday("2014/11/22")
}
\keyword{misc}

