% Generated by roxygen2 (4.0.2): do not edit by hand
\name{scores}
\alias{scores}
\title{What is the score?}
\usage{
scores(date)
}
\arguments{
\item{date}{The date you want to check (in YYYY-MM-DD format),
default vaule is system date}
}
\value{
Logical \code{TRUE} if \code{team} has a NHL game today,
\code{FALSE} otherwise
}
\description{
that takes date and returns a data.frame with
information about the games on date.
}
\details{
You know then problem: You're in your office writing R code and
suddenly have the urge to check whether your NHL team has a game today.
Before you know it you just wasted 15 minutes browsing the lastest
news on your favorite hockey webpage.
Suffer no more! You can now ask R directly, without tempting yourself
by firing up your web browser.
}
\note{
case in \code{team} is ignored
}
\examples{
scores()
scores("2014-11-22")
}

