% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{arena_team}
\alias{arena_team}
\title{Prices of 50,000 round cut diamonds.}
\format{A data frame with 30 rows and 4 variables:
\describe{
  \item{areana}{home city of the NHL team}
  \item{team}{NHL team name}
  \item{lat}{latitude of the home arena of the NHL team}
  \item{long}{longtitude of home arena of the NHL team}
}}
\usage{
arena_team
}
\description{
A dataset containing the team name of NHL, its home city and arena, latitude and longitude
of the arena.
}
\keyword{datasets}

