% Generated by roxygen2 (4.0.2): do not edit by hand
\name{check_date}
\alias{check_date}
\title{Is date a vilid date?}
\usage{
check_date(date = Sys.Date())
}
\arguments{
\item{NULL}{}
}
\value{
Logical. \code{TRUE} if the input is a vilde date,
\code{FALSE} otherwise
}
\description{
This function returns TRUE if the input is a vilde date
and FALSE otherwise
}
\details{
This function returns TRUE if the input is a vilde date
and FALSE otherwise
}
\keyword{misc}

