% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{gameday}
\alias{gameday}
\alias{gameday-package}
\title{The gameday Package}
\description{
Tell R to find out if your favourite NHL team is playing on a specific date,
or gather the scores of all teams on for a specific date. Stay up-to-date
on that hockey pool!
}

