% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{gameday}
\alias{gameday}
\alias{gameday-package}
\title{R package: gameday
This package is your personal assistant for checking NHL scores and matches}
\description{
R package: gameday
This package is your personal assistant for checking NHL scores and matches
}

