% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{gameday}
\alias{gameday}
\alias{gameday-package}
\title{My first R package: gameday}
\description{
Checks if NHL team has game day
}
\details{
This package allows the user to input an NHL team name, and checks to
see if the team is playing today.
}

