% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{gameday}
\alias{gameday}
\alias{gameday-package}
\title{My first R package: gameday}
\description{
This is so much fun
}
\details{
I look forward to writing really neat packages
so that I can have all my helper functions in one place
and maybe even share them.
}

