% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{gameday}
\alias{gameday}
\alias{gameday-package}
\title{My first R package: gameday}
\description{
Get NHL hockey statistics.
}
\details{
When living in Canada, you must:
\itemize{
\item Live hockey;
\item Breathe hockey;
\item Think about hockey all the time;
}

Or:
\itemize{
\item Be banished from civilization and inevitably perish in the tundra.
}

This package helps you to:
\itemize{
\item Prevent your banishment by providing basic hockey stats so you can
masquerade as a hockey lover. Go Canucks! Or whatever that Vancouver hockey
team is called...
}
}

