% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{gameday}
\alias{gameday}
\alias{gameday-package}
\title{gameday}
\description{
Get basic information about National Hockey League games from live.nhle.com.
}
\details{
When living in Canada, you must:
\itemize{
\item Live hockey;
\item Breathe hockey;
\item Think about hockey all the time;
}

Or:
\itemize{
\item Be banished from civilization and inevitably perish in the tundra.
}

The \code{gameday} helps you to:
\itemize{
\item Prevent your banishment by providing basic National Hockey League (NHL)
information so you can masquerade as a hockey lover.
}
}

