% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{gameday}
\alias{gameday}
\alias{gameday-package}
\title{gameday}
\description{
An RStudio interface for extracting NHL data from live.nhl.com
}
\details{
This package allows users to determine whether or not their NHL team
plays today or any other day of their choosing.
}

