% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{gameday}
\alias{gameday}
\alias{gameday-package}
\title{My first R package: gameday}
\description{
A one-line sentence on what this packages does
}
\details{
Single paragraph with more detail on the package.
This can span several lines.
}

