% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{gameday}
\alias{gameday}
\alias{gameday-package}
\title{My first R package: gameday}
\description{
A one-line summary of what this package is for
}
\details{
Then we can add a larger paragraph
that can span several
lines
}

