% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gday}
\alias{gday}
\title{Is it Gameday?}
\usage{
gday(team = "canucks", date = Sys.Date())
}
\arguments{
\item{team}{The team name you wish you search for. Defaults to \code{Canucks}}

\item{date}{Specified date to check whether the team is playing. Defaults to today's date}
}
\value{
\code{TRUE} if \code{team} has an NHL game on \code{date},
\code{FALSE} otherwise
}
\description{
This function returns TRUE if your NHL team plays today
and FALSE otherwise
}
\details{
You know then problem: You're in your office writing R code and
suddenly have the urge to check whether your NHL team has a game today.
Before you know it you just wasted 15 minutes browsing the lastest
news on your favorite hockey webpage.
Suffer no more! You can now ask R directly, without tempting yourself
by firing up your web browser.
}
\note{
case in \code{team} is ignored
}
\examples{
gday()
gday("canucks")
gday("Kings")
}
\keyword{misc}

