% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gday}
\alias{gday}
\title{Is it GameDay?}
\usage{
gday(team = "canucks")
}
\arguments{
\item{team}{}
}
\value{
Logical. \code{TRUE} if \code{team} has a NHL game today,
\code{FALSE} otherwise
}
\description{
Thus function returns TRUE if your NHL team plays today and false otherwise.
}
\details{
By default, I assume you are a Canucks fan.
}
\note{
case in \code{team} is ignored
}
\examples{
gday()
gday("Canadiens")
}
\keyword{misc}

