% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gday}
\alias{gday}
\title{Is it Gameday?}
\usage{
gday(team = "canucks", date = Sys.Date())
}
\arguments{
\item{team}{The name of your team; date}
}
\value{
Logical \code{TRUE} if \code{name} has an NHL game today
\code{FALSE} otherwise
}
\description{
This function returns TRUE is your NHL teams plays today and FALSE otherwise
}
\details{
You know then problem: You're in your office writing R code and
suddenly have the urge to check whether your NHL team has a game today.
Before you know it you just wasted 15 minutes browsing the lastest
news on your favorite hockey webpage.
Suffer no more! You can now ask R directly, without tempting yourself
by firing up your web browser.
}
\note{
case in \code{team} is ignored
}
\examples{
gday(team = "Canucks", date = "2014-11-23")
gday("Canucks")
gday("Bruins")
gday("New York")
gday(date = "2014-11-20")
}

