% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gday}
\alias{gday}
\title{Is it Gameday?}
\usage{
gday(team = "Canucks", gDate = Sys.Date())
}
\arguments{
\item{team}{The name of your team}

\item{gDate}{The day you want to check if the team plays}
}
\value{
Logical \code{TRUE} if \code{team} has a NHL game today,
\code{FALSE} otherwise
}
\description{
This function returns TRUE if your NHL team plays today and FALSE
otherwise
}
\note{
case in \code{team} is ignored. If gDate is not in format of
YYYY-mm-dd, then return an error.
}
\examples{
gday("canucks")
gday("Bruins")
}

