% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gday}
\alias{gday}
\title{Is it Gameday?}
\usage{
gday(team.name)
}
\arguments{
\item{team.name}{The name of your team}
}
\value{
Logical \code{TRUE} if \code{team.name} has a NHL game today,
\code{FALSE} otherwise
}
\description{
This function returns TRUE if your NHL team plays today and FALSE
otherwise
}
\note{
case in \code{team.name} is ignored
}
\examples{
gday("canucks")
gday("Bruins")
}

