% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gday}
\alias{gday}
\title{Is it Gameday?}
\usage{
gday(team = "canucks", date = Sys.Date())
}
\arguments{
\item{team}{Defaults to "canucks"}

\item{date}{Defaults to today}
}
\value{
Logical \code{TRUE} if \code{team} has an NHL game on a particular \code{date},
\code{FALSE} otherwise
}
\description{
This function returns TRUE if your NHL team plays on a particular day
and FALSE otherwise
}
\details{
You know the problem: You're in your office writing R code and
suddenly have the urge to check whether your NHL team has a game on a particular day.
Before you know it you just wasted 15 minutes browsing the lastest
news on your favorite hockey webpage.
Suffer no more! You can now ask R directly, without tempting yourself
by firing up your web browser.
}
\note{
case in \code{team} is ignored

city name gives the same results as team name
}
\examples{
gday()
gday("canadiens", "2014-11-17")
gday("Bruins")
gday("Vancouver")
}
\keyword{misc}

