% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gday}
\alias{gday}
\title{Is it Gameday?}
\usage{
gday(team = "canucks", date = Sys.Date())
}
\arguments{
\item{team}{The name of your team,}

\item{date}{The date the game occurred}
}
\value{
Logical \code{TRUE} if \code{team} has a NHL game on the date specified,
\code{FALSE} otherwise
}
\description{
This function returns TRUE if your NHL team plays on a date
and FALSE otherwise
}
\details{
You know then problem: You're in your office writing R code and
suddenly have the urge to check whether your NHL team has a game on a date.
Before you know it you just wasted 15 minutes browsing the lastest
news on your favorite hockey webpage.
Suffer no more! You can now ask R directly, without tempting yourself
by firing up your web browser.
}
\note{
case in \code{team} is ignored
}
\examples{
gday("canucks")
gday("Bruins")
}

