% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gday}
\alias{gday}
\title{Is it Gameday?}
\usage{
gday(team.name)
}
\arguments{
\item{team.name}{The name of your team}
}
\value{
Logical \code{TRUE} if \code{team.name} has a NHL game today,
\code{FALSE} otherwise
}
\description{
This function returns TRUE if you NHL team plays today and FALSE otherwise
}
\details{
You know then problem: You're in your office writing R code and
suddenly have the urge to check whether your NHL team has a game today.
Before you know it you just wasted 15 minutes browsing the lastest
news on your favorite hockey webpage.
Suffer no more! You can now ask R directly, without tempting yourself
by firing up your web browser.
}
\note{
case in \code{team.name} is ignored
}
\examples{
gday("canucks")
gday("Bruins")
}

