% Generated by roxygen2 (4.0.2): do not edit by hand
\name{internet_connection}
\alias{internet_connection}
\title{Is computer connected to internet?}
\usage{
internet_connection()
}
\arguments{
\item{NULL}{}
}
\value{
Logical. \code{TRUE} if the computer is connected to internet,
\code{FALSE} otherwise
}
\description{
This function returns TRUE if the computer is connected to internet
and FALSE otherwise
}
\details{
This function returns TRUE if the computer is connected to internet
and FALSE otherwise
}
\keyword{misc}

