% Generated by roxygen2 (4.0.2): do not edit by hand
\name{scores}
\alias{scores}
\title{What is the score?}
\usage{
scores(date = Sys.Date())
}
\arguments{
\item{date}{The date you want to check the scores (in YYYY-MM-DD format),
default vaule is system date}
}
\value{
A data.frame of listing home and away teams and their scores
}
\description{
Takes a date input and returns a data.frame with
information about the scores of all games on the given date.
}
\details{
Now you can check the scores of the NHL teams played/playing
on a given date, without the temptation to browse the web or
to watch the live stream...
}
\examples{
scores()
scores("2014-11-22")
}

