% Generated by roxygen2 (4.0.2): do not edit by hand
\name{scores}
\alias{scores}
\title{What is the score?}
\usage{
scores(date = Sys.Date())
}
\arguments{
\item{team}{}
}
\value{
a date frame with home team, away team, home_score and away_score.
}
\description{
This function returns the score of given date
}
\details{
You know then problem: You're in your office writing R code and
suddenly have the urge to check what the scores of NHL games are today.
Before you know it you just wasted 15 minutes browsing the lastest
news on your favorite hockey webpage.
Suffer no more! You can now ask R directly, without tempting yourself
by firing up your web browser.
}
\examples{
scores()
scores("2014-11-19")
}
\keyword{misc}

