% Generated by roxygen2 (4.0.2): do not edit by hand
\name{scores}
\alias{scores}
\title{Is it Gameday?}
\usage{
scores(date = Sys.Date())
}
\arguments{
\item{date}{Defaults to "Current date"}
}
\value{
\score of all games if games were played on the date
}
\description{
This function returns TRUE if your NHL team plays today
and FALSE otherwise
}
\details{
You know then problem: You're in your office writing R code and
suddenly have the urge to check whether your NHL team has a game today.
Before you know it you just wasted 15 minutes browsing the lastest
news on your favorite hockey webpage.
Suffer no more! You can now ask R directly, without tempting yourself
by firing up your web browser.
}
\note{
returns error if date does not follow yyyy-mm-dd format
}
\examples{
scores()
scores("2014-02-14")
scores("2014-2-14")
}
\keyword{misc}

