% Generated by roxygen2 (4.0.2): do not edit by hand
\name{scores}
\alias{scores}
\title{NHL final scores for a particular date}
\usage{
scores(date)
}
\arguments{
\item{date}{The specified date for which to look up scores for.}
}
\value{
A data frame with columns for the home and away team names, and for the home and away team scores.
}
\description{
This function returns a data frame with all scores on the specified date.
}
\details{
Allows the user to look up NHL final scores without going to the internet browser.
}
\note{
If the date is in the future, the columns for the scores will be empty.
}
\examples{
scores("2014-11-11")
scores("2015-01-01")
scores(Sys.Date())
}

