% Generated by roxygen2 (4.0.2): do not edit by hand
\name{scores}
\alias{scores}
\title{What's the Score?}
\usage{
scores(date = Sys.Date())
}
\arguments{
\item{date}{The date of the game}
}
\value{
Data.frame of scores
}
\description{
This function returns a dataframe of scores for a day's games, all within R.
}
\details{
Again, gameday assumes you are a Canucks fan. Sorry!
Imagine the new scenario. You have the hard choice of
staying on that awesome R coding session or watching your TV.
Suffer no more! You can simply ask R to give you the scores any day.
}
\note{
if \code{date} is not provided, games for today is checked
}
\examples{
scores()
scores("2014-11-22")
}

