% Generated by roxygen2 (4.0.2): do not edit by hand
\name{scores}
\alias{scores}
\title{NHL scores for a particular date}
\usage{
scores(date = Sys.Date())
}
\arguments{
\item{date}{The date you want to know the scores, format as "YYYY-MM-DD", default is \code{Sys.Date()}.}
}
\value{
a \code{data.frame} with variables: \code{home}, \code{away}, \code{home_score},
\code{away_score}, \code{game_status}, \code{home_shots}, \code{away_shots}.
}
\description{
This function returns a data.frame with information about the games on a particular
data you want to know.
}
\examples{
scores("2014-11-19")
}

