% Generated by roxygen2 (4.0.2): do not edit by hand
\name{scores}
\alias{scores}
\title{NHL scores for a particular day}
\usage{
scores(gDate)
}
\arguments{
\item{gDate}{The game day in YYYY-mm-dd format.}
}
\value{
data.frame The scores of every game played on the given day,
\code{gDate}. If no games are played on that day, then return an empty
\code{data.frame}
}
\description{
This function returns the NHL scores for every game played on a given
day.
}
\examples{
scores("2014-10-03")
scores("2014-07-10")
}

