% Generated by roxygen2 (4.0.2): do not edit by hand
\name{scores}
\alias{scores}
\title{What's the score?}
\usage{
scores(date = Sys.Date() - 1)
}
\arguments{
\item{date}{The date of the game, as "YYYY-MM-DD" (the default is yesterday's date)}
}
\value{
A dataframe with four variables: \code{home}, \code{away},
\code{home_score} and \code{away_score}, where each row contains the name of
the home team, their opposition (away team), and their respective scores.
}
\description{
This function returns the NHL game scores from a specified date.
}
\details{
The problem: you want to know the NHL scores from last night. Before you know
it you find yourself on google images trying to find the hottest NHL hockey
players...

Well fear no more, you can now ask R directly for the NHL scores from last
night (or any other date for that matter) without torturing your eyeballs!
}
\examples{
scores("2014-11-17")
scores("2013-12-11")
}

