% Generated by roxygen2 (4.0.2): do not edit by hand
\name{scores}
\alias{scores}
\title{What was the final score?}
\usage{
scores(date)
}
\arguments{
\item{date}{The specified date for which you want to know the scores.}
}
\value{
A data frame with the home and away team names and their scores for that day.
}
\description{
This function returns a data.frame with all scores from NHL games played on a certain date.
}
\details{
The user can look up NHL final scores for whichever day they please, without starting up a web browser.
}
\examples{
scores("2014-11-19")
scores("2013-10-01")
scores(Sys.Date())
}

