% Generated by roxygen2 (4.0.2): do not edit by hand
\name{scores}
\alias{scores}
\title{Is it Gameday?}
\usage{
scores(date = Sys.Date())
}
\arguments{
\item{date}{Defaults to today}
}
\value{
a data.frame about NHL game information on a particular \code{date}
}
\description{
This function returns a data.frame with information about the games on
a particular day.
}
\details{
You know the problem: You're in your office writing R code and
suddenly have the urge to check what are the game scores of the NHL teams
on a particular day.
Before you know it you just wasted 15 minutes browsing the lastest
news on your favorite hockey webpage.
Suffer no more! You can now ask R directly, without tempting yourself
by firing up your web browser.
}
\note{
case in \code{team} is ignored
}
\examples{
scores()
scores("2014-11-10")
}
\keyword{misc}

