% Generated by roxygen2 (4.0.2): do not edit by hand
\name{scores}
\alias{scores}
\title{What happened on this date?}
\usage{
scores(date = Sys.Date())
}
\arguments{
\item{date}{The date of interest}
}
\value{
\code{data.frame} containing data for the date supplied
}
\description{
This function returns info in a data.frame concerning games played on the date supplied.
}
\details{
You know the problem: You're in your office playing Trvial Pursuit: NHL Edition
when you suddenly have the urge to cheat and just get this game over with.
How will you do it? Is such a thing even possible? Yes!
You can now ask R directly, without wasting your time playing a game
dreamt up by only the most nefarious Canadian mind.
}
\examples{
scores("2014-11-02")
scores("2011-06-15")
}

