% Generated by roxygen2 (4.0.2): do not edit by hand
\name{scores}
\alias{scores}
\title{What are the final scores for the games at a particular date?}
\usage{
scores(date)
}
\arguments{
\item{\code{date}}{The specified date which you want to look the final scores
of all the games for.}
}
\value{
A data.frame with four columns. The first column shows the name of the home
team, the second column for the name of the away team. The third column gives
the score for the home team, while the score for the away team is given on the
forth column.
}
\description{
This function returns a data.frame with the final scores of the games on a
particular date.
}
\details{
You know the problem: You're in your office writing R code and
suddenly have the urge to check the final scores of the games on
a particular date.
Before you know it you just wasted 15 minutes browsing the lastest
news on your favorite hockey webpage.
Suffer no more! You can now ask R directly, without tempting yourself
by firing up your web browser.
}
\note{
The \code{date} must be entered as a string in the format: "yyyy-mm-dd".
The score columns will be empty if a future date is entered.
}
\examples{
scores(Sys.Date())
scores("2014-11-02")
scores("2014-12-31")
}

